requirements.txt
#Word2vec
gensim
jieba
numpy
pandas
matplotlib
scikit-learn
tqdm
opencc-python-reimplemented
bz2file